requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2==2.7.5
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.2.0
zipp==3.1.0
